iommu/vt-d: Add a helper to flush cache for updating present pasid entry
There are more paths that need to flush cache for present pasid entry as
pasid replacement is supported. Hence, add a helper for it. The major
code is in the intel_pasid_tear_down_entry(), so split this function into
a helper and call it in intel_pasid_tear_down_entry().

Signed-off-by: Yi Liu <yi.l.liu@intel.com>