iommufd: Allocate auto_domain with IOMMU_HWPT_ALLOC_PASID flag if device is PASID-capable
iommufd cannot predict if user will use pasid or not, but if user is
going to use it, the domains used by PASID-capable device should be
allocated with IOMMU_HWPT_ALLOC_PASID flag.

For domain allocated per userspace request, user should provide this
flag. While for the auto_domain, iommufd needs to make it.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>