iommufd: Mark PASID-compatible domain
AMD IOMMU requires attaching PASID-compatible domains to PASID-capable
devices. This includes the domains attached to RID and PASIDs. Related
discussions in link [1] and [2].  ARM has similar requirement but does
not need extra hint from iommufd, Intel does not have this requirement
but can live up with it. Hence, iommufd is going to enforce this
requirement as it's general requirement. Mark the PASID-capable domains
to prepare for adding this enforcement when PASID support is added.

[1] https://lore.kernel.org/linux-iommu/20240709182303.GK14050@ziepe.ca/
[2] https://lore.kernel.org/linux-iommu/20240822124433.GD3468552@ziepe.ca/

Signed-off-by: Yi Liu <yi.l.liu@intel.com>