iommufd: Enforce PASID-compatible domain for RID
Per the definition of IOMMU_HWPT_ALLOC_PASID, iommufd needs to enforce
the RID to use PASID-compatible domain if PASID has been attached, and
vice versa.

This enforcement requires a lock across the RID and PASID attach path,
the idev->igroup->lock is used as both RID and PASID path holds it.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>