vfio: Add vfio_copy_user_data()
This generalizes the logic of copying user data when the user struct
Have new fields introduced. The helpers can be used by the vfio uapis
that have the argsz and flags fields in the beginning 8 bytes.

As an example, the vfio_device_{at|de}tach_iommufd_pt paths are updated
to use the helpers.

The flags may be defined to mark a new field in the structure, reuse
reserved fields, or special handling of an existing field. The extended
size would differ for different flags. Each user API that wants to use
the generalized helpers should define an array to store the corresponding
extended sizes for each defined flag.

For example, we start out with the below, minsz is 12.

  struct vfio_foo_struct {
  	__u32   argsz;
  	__u32   flags;
  	__u32   pt_id;
  };

And then here it becomes:

  struct vfio_foo_struct {
  	__u32   argsz;
  	__u32   flags;
  #define VFIO_FOO_STRUCT_PASID   (1 << 0)
  	__u32   pt_id;
  	__u32   pasid;
  };

The array is { 16 }.

If the next flag is simply related to the processing of @pt_id and
doesn't require @pasid, then the extended size of the new flag is
12. The array become { 16, 12 }

  struct vfio_foo_struct {
  	__u32   argsz;
  	__u32   flags;
  #define VFIO_FOO_STRUCT_PASID   (1 << 0)
  #define VFIO_FOO_STRUCT_SPECICAL_PTID   (1 << 1)
  	__u32   pt_id;
  	__u32   pasid;
  };

Similarly, rather than adding new field, we might have reused a previously
reserved field, for instance what if we already expanded the structure
as the below, array is already { 24 }.

  struct vfio_foo_struct {
  	__u32   argsz;
  	__u32   flags;
  #define VFIO_FOO_STRUCT_XXX     (1 << 0)
  	__u32   pt_id;
  	__u32   reserved;
  	__u64   xxx;
  };

If we then want to add @pasid, we might really prefer to take advantage
of that reserved field and the array becomes { 24, 16 }.

  struct vfio_foo_struct {
  	__u32   argsz;
  	__u32   flags;
  #define VFIO_FOO_STRUCT_XXX     (1 << 0)
  #define VFIO_FOO_STRUCT_PASID   (1 << 1)
  	__u32   pt_id;
  	__u32   pasid;
  	__u64   xxx;
  };

Suggested-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>