vfio: Generalize the logic of copying user data for structures with extension
The struct vfio_device_attach_iommufd_pt and struct vfio_device_detach_iommu
have been extended to include a @pasid field marked by a flag in the @flags
field. The attach/detach path have the shared logic to copy user data. So
it makes a lot sense to generalize it.

This generalized code can be used by the vfio uapis that has the argsz and
flags fields in the beginning 8 bytes. The flags may be defined to mark a
new field in the structure, reusing a reserved field, speical handling of
an existing field. Hence, the way to calculate the extended size would differ.
For each user API that wants to make use of the generalized helpers, it should
define an array to include the corresponding extended size for the flags
defined.

For example, we start out with:

struct vfio_device_attach_iommufd_pt {
        __u32   argsz;
        __u32   flags;
        __u32   pt_id;
};

And then here it becomes:

struct vfio_device_attach_iommufd_pt {
	__u32	argsz;
	__u32	flags;
	__u32	pt_id;
	__u32	pasid;
};

The array is { 16 }.

If the next flag is simply related to the processing of @pt_id and
doesn't require @pasid, then the extended size of the new flag is
12. The array become { 16, 12 }

Similarly, rather than pasid we might have reused a previously
reserved field, for instance what if we already expanded the structure
as:

struct vfio_device_attach_iommufd_pt {
	__u32	argsz;
	__u32	flags;
	__u32	pt_id;
	__u32	reserved;
	__u64	foo;
};

If we then want to add @pasid, we might really prefer to take advantage
of that reserved field and the array becomes { 24, 16 }.

Suggested-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>