iommu: Swap the order of setting group->pasid_array and calling attach op of iommu drivers
The current implementation stores group->pasid_array before the underlying
iommu driver has successfully set the new domain. This can lead to issues
where PRIs are received on the new domain before the attach operation is
completed. This patch swaps the order of operations to ensure that the
domain is set in the iommu driver before updating group->pasid_array.

Suggested-by: Jason Gunthorpe <jgg@nvidia.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>